refactor(helix-tui): split `text.rs` into submodules
